Simplify autorevision setup
Remove autorevision.cache file as autorevision is run every
time a build is started. Autorevision is included in repo so
a build should not fail on another machine.

For CLI Target a swift file is generated which is included in
the Xcode project (file browser)

For the GUI Target a header file is generated which is *NOT*
included in Xcode (file browser) but is inside Supporing Files.
info.plist preprocessor will replace the values during compilation.